Fix test_service_state() so it always returns 1 when it's not in that state.
svn path=/branches/baselayout-1_12/; revision=1568

+++ sbin/rc-services.sh
@@ -540,12 +540,8 @@ test_service_state() {
 	# Service is in the state requested
 	[[ -L ${f} ]] && return 0
 	
-	if [[ ! -e ${f} ]]; then
-		rm -f "${f}"
-		return 1
-	fi
-	
-	return 0
+	[[ ! -e ${f} ]] && rm -f "${f}"
+	return 1
 }
 
 # bool service_starting(service)
